Updates to web server
Expose static directory, where result files should be written
MakeMosaic should now write the result image to static and return
the filename. The router will then respond to the user's request
with a redirect to the image file.

Refactored the database initialization code to load the image index
from a json file instead of processing each image every time the
web server is initialized.

Parse the url for the source image from the user's request.  It
should be in the 'img' query parameter.